Fixed bug about modal edit dialog.
What happened is that if we left an edit modal dialog open, or open an
edit dialog and then not clicking that save button, the flag
_isEditingUser was left being true. That caused a bug which didn't
permit us to add a new Employee until we clicked two times the save
button.
Modal_Activated and Modal_Deactivated solves this by subscribing the
Window.Activated and Window.Deactivated events of the modal.