Properties to Login SuccessMessage.
This message will be used to show it in the LoginView. For example, if
the user forgets the password, after requesting a reminder the view will
go to the Login and there it will show a message that tells the password
reminder was sent successfull.
A similar thing will happen when the first admin registers.