Adding information icon.
That makes clear the usage of "Active" employee checkbox.
Also, hidding the checkbox of "Active" employee, since the concept of
"Firing an employee" when the employee is being created (Hired) doesn't
makes sense.
So, it hides this option when creating a new employee, but it shows when
editing an existing one.