Fixed modal showing when editing non admin employee.
This happened if we opened the edit view of the last administrator,
close the window, and then open an edit in a normal user. It showed the
"can't edit user type because it's the last admin." due to the UserType
set accessor validation.

Also, changed the literal string to a resource one.