Fixed bug; "User to Admin to User" Combobox.
We got an modal error saying that we can't change the Administrator to a
Employee user type. This happened by doing the following:
If we have only one administrator in the collection, then go edit a
normal user and change the User type's combobox to "Administrator" and
then again to "Employee" that's when the error happened due to our
validations.